requirements.txt
# PyTorch must be installed separately — see README Installation section
# https://pytorch.org/get-started/locally/
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0
Pillow>=9.5.0
huggingface_hub>=0.20.0
